+++ src/ui/popover/wasshop/wasshop.dialog.ts
@@ -1,5 +1,6 @@
 import { Component, Inject, ChangeDetectorRef } from '@angular/core';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
+import { Observable } from 'rxjs/Observable';
 import { UserService } from '../../../user.service';
 import { User } from '../../../app.models';
 import { WasUp } from '../../../ui/popover/wasup/wasup.dialog';
@@ -40,8 +41,6 @@ export class WasShop {
   }
   /**
    * Cancel/close the dialog
-   *
-   * @memberof WasShop
    */
   onNoClick(): void {
     this.dialogRef.close();